misc: refactor main
