+++ pom.xml
@@ -142,7 +142,7 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <netbeans.compile.on.save>none</netbeans.compile.on.save>
-        <exec.mainClass>org.apidesign.WebSocket</exec.mainClass>
+        <exec.mainClass>org.apidesign.polyfill.websocket.Main</exec.mainClass>
         <inspectPort></inspectPort>
         <exec.vmArgs>-DinspectPort=${inspectPort}</exec.vmArgs>
         <versions.helidon>4.0.2</versions.helidon>
