+++ src/test/java/org/apidesign/polyfill/WebSocketPolyfillTest.java
@@ -1,9 +1,10 @@
-package org.apidesign;
+package org.apidesign.polyfill;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
+import org.apidesign.polyfill.websocket.WebSocketPolyfill;
 import org.graalvm.polyglot.Context;
 import org.graalvm.polyglot.Source;
 import org.graalvm.polyglot.io.IOAccess;
@@ -29,10 +30,11 @@ public class WebSocketPolyfillTest {
             b.option("inspect", ":" + chromePort);
         }
         executor = Executors.newSingleThreadExecutor();
+        var webSocketPolyfill = new WebSocketPolyfill(executor);
         ctx = CompletableFuture
                 .supplyAsync(() -> {
                     var context = b.build();
-                    WebSocket.initializePolyfill(context, executor);
+                    webSocketPolyfill.initialize(context);
                     return context;
                 }, executor)
                 .get();
